use stream event in test
